Uploading an incomplete character controller
Cursor needs work in order to function properly but that should only take a few hours. And the character controller does not yet collide with walls so there are currently no ways to stop them flying around. Other than that it works like the editor movement and I think that should be ok for now.